pyproject.toml
[project]
name = "libdb"
version = "0.1.0"
description = "A basic book inventory tracker"
authors = [
    {name = "Dom Chester",email = "dom.chester@me.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "flask (>=3.1.1,<4.0.0)",
    "sqlalchemy (>=2.0.41,<3.0.0)",
    "flask-sqlalchemy (>=3.1.1,<4.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "flask-wtf (>=1.2.2,<2.0.0)",
    "wtforms-sqlalchemy (>=0.4.2,<0.5.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "flask-alembic (>=3.1.1,<4.0.0)"
]

[tool.poetry.group.dev.dependencies]
taskipy = "^1.14.1"
flake8 = "^7.3.0"
flake8-blind-except = "^0.2.1"
flake8-bugbear = "^24.12.12"
flake8-builtins = "^2.5.0"
flake8-debugger = "^4.1.2"
flake8-docstrings = "^1.7.0"
flake8-eradicate = "^1.5.0"
flake8-simplify = "^0.22.0"
pytest = "^8.4.1"
pytest-cov = "^6.2.1"
bandit = "^1.8.5"
mypy = "^1.16.1"
types-wtforms = "^3.2.1.20250602"
types-flask = "^1.1.6"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.taskipy.tasks]
start = "flask --app libdb run --debug"
pytest = "pytest --cov=libdb tests/ --cov-report term:skip-covered"
flake8 = "flake8 libdb"
bandit = "bandit -c pyproject.toml -r libdb"
mypy = "mypy libdb"
db_revision = "flask db revision"
db_upgrade = "flask db upgrade"

[tool.isort]
profile = "black"
line_length = 88

[tool.black]
line-length = 88
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  # The following are specific to Black, you probably don't want those.
  | blib2to3
  | tests/data
  | profiling
)/
'''

[tool.typos]
default.extend-words = { }

[[tool.mypy.overrides]]
module = ["flask_wtf.*"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["wtforms_sqlalchemy.*"]
ignore_missing_imports = true